tests/conftest.py
"""Pytest configuration for tui-chatbot tests."""

import asyncio
import sys
from pathlib import Path

# Add src to path
sys.path.insert(0, str(Path(__file__).parent.parent / "src"))

import pytest


@pytest.fixture
def event_loop():
    """Create a new event loop for each test."""
    loop = asyncio.new_event_loop()
    yield loop
    loop.close()


@pytest.fixture
def async_event_loop():
    """Get the current event loop for async tests."""
    return asyncio.get_event_loop()
