pyproject.toml
[project]
name = "tui-chatbot"
version = "0.1.0"
description = "A modern TUI chatbot with streaming output and TPS statistics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "pydantic>=2.0.0",
]

[project.scripts]
chat = "tui_chatbot.main:main"
